1971/01/21 - Issue #54 1971 missing setlists
I'm confused about the setlist for 1/21.  Both copies on archive.org appear to not include Uncle John's Band and both appear to have some filler from some other show (ah, the days of 90 minute tapes). I went with Uncle John's Band not as an encore as per The SetList Program, but I'm not clear if it's an encore.  There's a loud crowd response after China-Rider, which makes me think it was an encore.  I'm going to reach out to a facebook group and see if someone can help me figure it out for sure.   @jefmsmit - open to suggestions.